requirements.txt
APScheduler==3.6.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1.7
django-allauth==0.39.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-cors-headers==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
environ==1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
#mysqlclient==1.4.2.post1
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.3

